Set proper file label for dri dir
It seems all the libraries under /vendor/lib64/dri
need to be accessed by some core domains. Set
same_process_hal_file label for the whole dir of
/vendor/lib64/dri to avoid too many items in the
file_contexts file. And this can also fix an issue
that if a dri library is pushed to dri dir after
remount and the device cannot boot up

Tracked-On: OAM-123676
Signed-off-by: jizhenlo <zhenlong.z.ji@intel.com>